publish: kinesis: add tests for stream's existence check
